fix: #31 바텀 탭 삭제
